Moved code to new files.
