+++ src/tests/NeXusDatasetTests.cpp
@@ -8,13 +8,12 @@
 // Screaming Udder!                              https://esss.se
 
 #include "NeXusDataset.h"
-#include "schemas/senv/FastSampleEnvironmentWriter.h"
 #include <gtest/gtest.h>
 #include <h5cpp/dataspace/simple.hpp>
 #include <h5cpp/datatype/type_trait.hpp>
 #include <h5cpp/hdf5.hpp>
 
-class DatasetCreation : public ::testing::Test {
+class NeXusDatasetCreation : public ::testing::Test {
 public:
   void SetUp() override {
     File = hdf5::file::create(TestFileName, hdf5::file::AccessFlags::TRUNCATE);
@@ -27,283 +26,9 @@ public:
   hdf5::node::Group RootGroup;
 };
 
-TEST_F(DatasetCreation, NDArrOpen) {
-  {
-    size_t ChunkSize{256};
-    NeXusDataset::MultiDimDataset<int> ADValues(
-        RootGroup, NeXusDataset::Mode::Create, {10, 10}, {ChunkSize});
-  }
-  EXPECT_NO_THROW(NeXusDataset::MultiDimDataset<int> ReOpened(
-      RootGroup, NeXusDataset::Mode::Open));
-}
-
-TEST_F(DatasetCreation, NDArrOpenAlt2) {
-  NeXusDataset::MultiDimDataset<int> ADValues(
-      RootGroup, NeXusDataset::Mode::Create, {10, 10}, {5, 6, 4, 3});
-  auto ChunkDims = ADValues.creation_list().chunk();
-  EXPECT_EQ(ChunkDims, (hdf5::Dimensions{1024, 10, 10}));
-}
-
-TEST_F(DatasetCreation, NDArrOpenAlt3) {
-  NeXusDataset::MultiDimDataset<int> ADValues(
-      RootGroup, NeXusDataset::Mode::Create, {10, 10}, {10, 10, 10});
-  auto ChunkDims = ADValues.creation_list().chunk();
-  EXPECT_EQ(ChunkDims, (hdf5::Dimensions{10, 10, 10}));
-}
-
-TEST_F(DatasetCreation, NDArrOpenAlt) {
-  size_t ChunkSize{256};
-  {
-    NeXusDataset::MultiDimDataset<int> ADValues(
-        RootGroup, NeXusDataset::Mode::Create, {10, 10}, {ChunkSize});
-  }
-  EXPECT_NO_THROW(NeXusDataset::MultiDimDataset<int>(
-      RootGroup, NeXusDataset::Mode::Open, {10, 10}, {ChunkSize}));
-}
-
-TEST_F(DatasetCreation, NDArrCreateFail) {
-  EXPECT_THROW(NeXusDataset::MultiDimDataset<int> ReOpened(
-                   RootGroup, NeXusDataset::Mode::Create),
-               std::runtime_error);
-}
-
-TEST_F(DatasetCreation, NDArrConstructorFail) {
-  EXPECT_THROW(NeXusDataset::MultiDimDataset<int> ReOpened(
-                   RootGroup, NeXusDataset::Mode(-43536)),
-               std::runtime_error);
-}
-
-TEST_F(DatasetCreation, NDArrCreationMaxSize) {
-  hdf5::Dimensions DatasetDimensions{10, 10};
-  {
-    size_t ChunkSize{256};
-    NeXusDataset::MultiDimDataset<int> ADValues(
-        RootGroup, NeXusDataset::Mode::Create, DatasetDimensions, {ChunkSize});
-  }
-  NeXusDataset::MultiDimDataset<int> ReOpened(RootGroup,
-                                              NeXusDataset::Mode::Open);
-  auto DataSpace = ReOpened.dataspace();
-  hdf5::dataspace::Simple SomeSpace(DataSpace);
-  auto MaxDims = SomeSpace.maximum_dimensions();
-  for (auto i : MaxDims) {
-    EXPECT_EQ(
-        i,
-        H5S_UNLIMITED); // Due to bug in h5cpp I am not using Simple::UNLIMITED
-  }
-}
-
-TEST_F(DatasetCreation, NDArrCreationArrSize) {
-  hdf5::Dimensions DatasetDimensions{10, 10};
-  {
-    size_t ChunkSize{256};
-    NeXusDataset::MultiDimDataset<int> ADValues(
-        RootGroup, NeXusDataset::Mode::Create, DatasetDimensions, {ChunkSize});
-  }
-  NeXusDataset::MultiDimDataset<int> ReOpened(RootGroup,
-                                              NeXusDataset::Mode::Open);
-  auto DataSpace = ReOpened.dataspace();
-  hdf5::dataspace::Simple SomeSpace(DataSpace);
-  auto NewDims = SomeSpace.current_dimensions();
-  DatasetDimensions.insert(DatasetDimensions.begin(), 0);
-  EXPECT_EQ(DatasetDimensions, NewDims);
-}
-
-TEST_F(DatasetCreation, NDArrCreationChunkSize1) {
-  hdf5::Dimensions DatasetDimensions{10, 10};
-  size_t ChunkSize{256};
-  {
-    NeXusDataset::MultiDimDataset<int> ADValues(
-        RootGroup, NeXusDataset::Mode::Create, DatasetDimensions, {ChunkSize});
-  }
-  NeXusDataset::MultiDimDataset<int> ReOpened(RootGroup,
-                                              NeXusDataset::Mode::Open);
-  auto CreationProperties = ReOpened.creation_list();
-  auto ChunkDims = CreationProperties.chunk();
-  DatasetDimensions.insert(DatasetDimensions.begin(), ChunkSize);
-  EXPECT_EQ(ChunkDims, DatasetDimensions);
-}
-
-TEST_F(DatasetCreation, NDArrCreationChunkSize2) {
-  hdf5::Dimensions DatasetDimensions{10, 10};
-  hdf5::Dimensions ChunkSize = {100, 20, 20};
-  {
-    NeXusDataset::MultiDimDataset<int> ADValues(
-        RootGroup, NeXusDataset::Mode::Create, DatasetDimensions, {ChunkSize});
-  }
-  NeXusDataset::MultiDimDataset<int> ReOpened(RootGroup,
-                                              NeXusDataset::Mode::Open);
-  auto CreationProperties = ReOpened.creation_list();
-  auto ChunkDims = CreationProperties.chunk();
-  EXPECT_EQ(ChunkDims, ChunkSize);
-}
-
-TEST_F(DatasetCreation, NDArrCreationChunkSize3) {
-  hdf5::Dimensions DatasetDimensions{10, 10};
-  hdf5::Dimensions ChunkSize = {};
-  {
-    NeXusDataset::MultiDimDataset<int> ADValues(
-        RootGroup, NeXusDataset::Mode::Create, DatasetDimensions, {ChunkSize});
-  }
-  NeXusDataset::MultiDimDataset<int> ReOpened(RootGroup,
-                                              NeXusDataset::Mode::Open);
-  auto CreationProperties = ReOpened.creation_list();
-  auto ChunkDims = CreationProperties.chunk();
-  DatasetDimensions.insert(DatasetDimensions.begin(), 1024);
-  EXPECT_EQ(ChunkDims, DatasetDimensions);
-}
-
-TEST_F(DatasetCreation, NDArrCreationChunkSize4) {
-  hdf5::Dimensions DatasetDimensions{10, 10};
-  hdf5::Dimensions ChunkSize = {1, 2, 3, 4, 5};
-  {
-    NeXusDataset::MultiDimDataset<int> ADValues(
-        RootGroup, NeXusDataset::Mode::Create, DatasetDimensions, {ChunkSize});
-  }
-  NeXusDataset::MultiDimDataset<int> ReOpened(RootGroup,
-                                              NeXusDataset::Mode::Open);
-  auto CreationProperties = ReOpened.creation_list();
-  auto ChunkDims = CreationProperties.chunk();
-  DatasetDimensions.insert(DatasetDimensions.begin(), 1024);
-  EXPECT_EQ(ChunkDims, DatasetDimensions);
-}
-
-TEST_F(DatasetCreation, NDArrAppendSameSize) {
-  hdf5::Dimensions DatasetDimensions{2, 2};
-  NeXusDataset::MultiDimDataset<int> Dataset(
-      RootGroup, NeXusDataset::Mode::Create, DatasetDimensions, {});
-  std::vector<int> TestData{2, 4, 6, 8};
-  Dataset.appendArray(TestData, DatasetDimensions);
-  DatasetDimensions.insert(DatasetDimensions.begin(), 1);
-  EXPECT_EQ(DatasetDimensions, Dataset.get_extent());
-  std::vector<int> StoredData(TestData.size());
-  Dataset.read(StoredData);
-  EXPECT_EQ(TestData, StoredData);
-}
-
-TEST_F(DatasetCreation, NDArrAppendAnotherType) {
-  hdf5::Dimensions DatasetDimensions{2, 2};
-  NeXusDataset::MultiDimDataset<double> Dataset(
-      RootGroup, NeXusDataset::Mode::Create, DatasetDimensions, {});
-  std::vector<int> TestData{2, 4, 6, 8};
-  std::vector<double> CompareData{2, 4, 6, 8};
-  Dataset.appendArray(TestData, DatasetDimensions);
-  std::vector<double> StoredData(TestData.size());
-  Dataset.read(StoredData);
-  EXPECT_EQ(CompareData, StoredData);
-}
-
-TEST_F(DatasetCreation, NDArrAppendSmallerSize1) {
-  hdf5::Dimensions DatasetDimensions{2, 2};
-  NeXusDataset::MultiDimDataset<int> Dataset(
-      RootGroup, NeXusDataset::Mode::Create, DatasetDimensions, {});
-  std::vector<int> TestData{6, 8};
-  Dataset.appendArray(TestData, {1, 2});
-  DatasetDimensions.insert(DatasetDimensions.begin(), 1);
-  EXPECT_EQ(DatasetDimensions, Dataset.get_extent());
-  TestData = std::vector<int>{6, 8, 0, 0};
-  std::vector<int> StoredData(TestData.size());
-  Dataset.read(StoredData);
-  EXPECT_EQ(TestData, StoredData);
-}
-
-TEST_F(DatasetCreation, NDArrAppendWrongRank) {
-  hdf5::Dimensions DatasetDimensions{2, 2};
-  NeXusDataset::MultiDimDataset<int> Dataset(
-      RootGroup, NeXusDataset::Mode::Create, DatasetDimensions, {});
-  std::vector<int> TestData{6, 8};
-  EXPECT_THROW(Dataset.appendArray(TestData, {2}), std::runtime_error);
-}
-
-TEST_F(DatasetCreation, NDArrAppendSmallerSize2) {
-  hdf5::Dimensions DatasetDimensions{2, 2};
-  NeXusDataset::MultiDimDataset<int> Dataset(
-      RootGroup, NeXusDataset::Mode::Create, DatasetDimensions, {});
-  std::vector<int> TestData{6, 8};
-  Dataset.appendArray(TestData, {2, 1});
-  DatasetDimensions.insert(DatasetDimensions.begin(), 1);
-  EXPECT_EQ(DatasetDimensions, Dataset.get_extent());
-  TestData = std::vector<int>{6, 0, 8, 0};
-  std::vector<int> StoredData(TestData.size());
-  Dataset.read(StoredData);
-  EXPECT_EQ(TestData, StoredData);
-}
-
-TEST_F(DatasetCreation, NDArrAppendBiggerSize1) {
-  hdf5::Dimensions DatasetDimensions{2, 2};
-  NeXusDataset::MultiDimDataset<int> Dataset(
-      RootGroup, NeXusDataset::Mode::Create, DatasetDimensions, {});
-  std::vector<int> TestData{6, 8, 10, 12, 14, 16};
-  Dataset.appendArray(TestData, {2, 3});
-  DatasetDimensions.insert(DatasetDimensions.begin(), 1);
-  DatasetDimensions[2] = 3;
-  EXPECT_EQ(DatasetDimensions, Dataset.get_extent());
-  std::vector<int> StoredData(TestData.size());
-  Dataset.read(StoredData);
-  EXPECT_EQ(TestData, StoredData);
-}
-
-TEST_F(DatasetCreation, NDArrAppendWrongDimensions) {
-  hdf5::Dimensions DatasetDimensions{2, 2};
-  NeXusDataset::MultiDimDataset<int> Dataset(
-      RootGroup, NeXusDataset::Mode::Create, DatasetDimensions, {});
-  std::vector<int> TestData{6, 8};
-  EXPECT_THROW(Dataset.appendArray(TestData, {2}), std::runtime_error);
-}
-
-TEST_F(DatasetCreation, AppendDataOnce) {
-  int ChunkSize = 256;
-  std::array<const std::uint16_t, 4> SomeData{{0, 1, 2, 3}};
-  NeXusDataset::ExtensibleDataset<std::uint16_t> TestDataset(
-      RootGroup, "SomeDataset", NeXusDataset::Mode::Create, ChunkSize);
-  TestDataset.appendArray(SomeData);
-  auto DataspaceSize = TestDataset.dataspace().size();
-  EXPECT_EQ(static_cast<uint64_t>(DataspaceSize), SomeData.size());
-  std::vector<std::uint16_t> Buffer(DataspaceSize);
-  TestDataset.read(Buffer);
-  for (int i = 0; i < DataspaceSize; i++) {
-    ASSERT_EQ(Buffer.at(i), SomeData.at(i));
-  }
-}
-
-TEST_F(DatasetCreation, AppendDataTwice) {
-  int ChunkSize = 256;
-  std::array<const std::uint16_t, 4> SomeData{{0, 1, 2, 3}};
-  NeXusDataset::ExtensibleDataset<std::uint16_t> TestDataset(
-      RootGroup, "SomeDataset", NeXusDataset::Mode::Create, ChunkSize);
-  TestDataset.appendArray(SomeData);
-  TestDataset.appendArray(SomeData);
-  auto DataspaceSize = TestDataset.dataspace().size();
-  EXPECT_EQ(static_cast<uint64_t>(DataspaceSize), SomeData.size() * 2);
-  std::vector<std::uint16_t> Buffer(DataspaceSize);
-  TestDataset.read(Buffer);
-  for (int i = 0; i < DataspaceSize; i++) {
-    ASSERT_EQ(Buffer.at(i), SomeData.at(i % SomeData.size()))
-        << "Failed at i = " << i;
-  }
-}
-
-TEST_F(DatasetCreation, AppendArrayAdpaterDataTwice) {
-  int ChunkSize = 256;
-  std::array<const std::uint16_t, 4> SomeData{{0, 1, 2, 3}};
-  NeXusDataset::ExtensibleDataset<std::uint16_t> TestDataset(
-      RootGroup, "SomeDataset", NeXusDataset::Mode::Create, ChunkSize);
-  ArrayAdapter<const std::uint16_t> TempAdapter{
-      SomeData.data(), static_cast<size_t>(SomeData.size())};
-  TestDataset.appendArray(TempAdapter);
-  TestDataset.appendArray(TempAdapter);
-  auto DataspaceSize = TestDataset.dataspace().size();
-  EXPECT_EQ(static_cast<uint64_t>(DataspaceSize), SomeData.size() * 2);
-  std::vector<std::uint16_t> Buffer(DataspaceSize);
-  TestDataset.read(Buffer);
-  for (int i = 0; i < DataspaceSize; i++) {
-    ASSERT_EQ(Buffer.at(i), SomeData.at(i % SomeData.size()))
-        << "Failed at i = " << i;
-  }
-}
-
 //--------------------------------------------------
 
-TEST_F(DatasetCreation, RawValueDefaultCreation) {
+TEST_F(NeXusDatasetCreation, RawValueDefaultCreation) {
   size_t ChunkSize = 256;
   {
     NeXusDataset::RawValue ADCValues(RootGroup, NeXusDataset::Mode::Create,
@@ -318,14 +43,14 @@ TEST_F(DatasetCreation, RawValueDefaultCreation) {
   EXPECT_EQ(hdf5::datatype::create<std::uint16_t>(), TestDataset.datatype());
 }
 
-TEST_F(DatasetCreation, RawValueConstructorFail) {
+TEST_F(NeXusDatasetCreation, RawValueConstructorFail) {
   size_t ChunkSize = 256;
   EXPECT_THROW(NeXusDataset::RawValue(RootGroup, NeXusDataset::Mode(-1247832),
                                       ChunkSize),
                std::runtime_error);
 }
 
-TEST_F(DatasetCreation, RawValueReOpen) {
+TEST_F(NeXusDatasetCreation, RawValueReOpen) {
   {
     size_t ChunkSize{256};
     NeXusDataset::RawValue ADCValues(RootGroup, NeXusDataset::Mode::Create,
@@ -335,7 +60,7 @@ TEST_F(DatasetCreation, RawValueReOpen) {
       NeXusDataset::RawValue ReOpened(RootGroup, NeXusDataset::Mode::Open));
 }
 
-TEST_F(DatasetCreation, RawValueThrowOnExists) {
+TEST_F(NeXusDatasetCreation, RawValueThrowOnExists) {
   size_t ChunkSize = 256;
   {
     NeXusDataset::RawValue ADCValues(RootGroup, NeXusDataset::Mode::Create,
@@ -348,7 +73,7 @@ TEST_F(DatasetCreation, RawValueThrowOnExists) {
 
 //--------------------------------------------------
 
-TEST_F(DatasetCreation, TimeDefaultCreation) {
+TEST_F(NeXusDatasetCreation, TimeDefaultCreation) {
   size_t ChunkSize = 256;
   {
     NeXusDataset::Time Timestamps(RootGroup, NeXusDataset::Mode::Create,
@@ -381,7 +106,7 @@ TEST_F(DatasetCreation, TimeDefaultCreation) {
   EXPECT_TRUE(FoundUnitAttr);
 }
 
-TEST_F(DatasetCreation, TimeReOpen) {
+TEST_F(NeXusDatasetCreation, TimeReOpen) {
   {
     size_t ChunkSize{256};
     NeXusDataset::Time Timestamps(RootGroup, NeXusDataset::Mode::Create,
@@ -391,7 +116,7 @@ TEST_F(DatasetCreation, TimeReOpen) {
       NeXusDataset::Time ReOpened(RootGroup, NeXusDataset::Mode::Open));
 }
 
-TEST_F(DatasetCreation, TimeThrowOnExists) {
+TEST_F(NeXusDatasetCreation, TimeThrowOnExists) {
   size_t ChunkSize = 256;
   {
     NeXusDataset::Time Timestamps(RootGroup, NeXusDataset::Mode::Create,
@@ -404,7 +129,7 @@ TEST_F(DatasetCreation, TimeThrowOnExists) {
 
 //--------------------------------------------------
 
-TEST_F(DatasetCreation, CueIndexDefaultCreation) {
+TEST_F(NeXusDatasetCreation, CueIndexDefaultCreation) {
   size_t ChunkSize = 256;
   {
     NeXusDataset::CueIndex Cue(RootGroup, NeXusDataset::Mode::Create,
@@ -419,7 +144,7 @@ TEST_F(DatasetCreation, CueIndexDefaultCreation) {
   EXPECT_EQ(hdf5::datatype::create<std::uint32_t>(), TestDataset.datatype());
 }
 
-TEST_F(DatasetCreation, CueIndexReOpen) {
+TEST_F(NeXusDatasetCreation, CueIndexReOpen) {
   {
     size_t ChunkSize{256};
     NeXusDataset::CueIndex Cue(RootGroup, NeXusDataset::Mode::Create,
@@ -429,7 +154,7 @@ TEST_F(DatasetCreation, CueIndexReOpen) {
       NeXusDataset::CueIndex ReOpened(RootGroup, NeXusDataset::Mode::Open));
 }
 
-TEST_F(DatasetCreation, CueIndexThrowOnExists) {
+TEST_F(NeXusDatasetCreation, CueIndexThrowOnExists) {
   size_t ChunkSize = 256;
   {
     NeXusDataset::CueIndex Cue(RootGroup, NeXusDataset::Mode::Create,
@@ -442,7 +167,7 @@ TEST_F(DatasetCreation, CueIndexThrowOnExists) {
 
 //--------------------------------------------------
 
-TEST_F(DatasetCreation, CueTimestampZeroDefaultCreation) {
+TEST_F(NeXusDatasetCreation, CueTimestampZeroDefaultCreation) {
   size_t ChunkSize = 256;
   {
     NeXusDataset::CueTimestampZero Cue(RootGroup, NeXusDataset::Mode::Create,
@@ -475,7 +200,7 @@ TEST_F(DatasetCreation, CueTimestampZeroDefaultCreation) {
   EXPECT_TRUE(FoundUnitAttr);
 }
 
-TEST_F(DatasetCreation, CueTimestampZeroReOpen) {
+TEST_F(NeXusDatasetCreation, CueTimestampZeroReOpen) {
   {
     size_t ChunkSize{256};
     NeXusDataset::CueTimestampZero Cue(RootGroup, NeXusDataset::Mode::Create,
@@ -485,7 +210,7 @@ TEST_F(DatasetCreation, CueTimestampZeroReOpen) {
       RootGroup, NeXusDataset::Mode::Open));
 }
 
-TEST_F(DatasetCreation, CueTimestampZeroThrowOnExists) {
+TEST_F(NeXusDatasetCreation, CueTimestampZeroThrowOnExists) {
   size_t ChunkSize = 256;
   {
     NeXusDataset::CueTimestampZero Cue(RootGroup, NeXusDataset::Mode::Create,
@@ -496,99 +221,172 @@ TEST_F(DatasetCreation, CueTimestampZeroThrowOnExists) {
                std::runtime_error);
 }
 
-TEST_F(DatasetCreation, StringDatasetDefaultCreation) {
-  std::string DatasetName{"SomeName"};
-  size_t StringLength{24};
-  size_t ChunkSize{511};
-  {
-    NeXusDataset::FixedSizeString Strings(RootGroup, DatasetName,
-                                          NeXusDataset::Mode::Create,
-                                          StringLength, ChunkSize);
-  }
+//--------------------------------------------------
 
-  ASSERT_TRUE(RootGroup.has_dataset(DatasetName));
-  hdf5::node::Dataset TestDataset = RootGroup.get_dataset(DatasetName);
+TEST_F(NeXusDatasetCreation, EventIdDefaultCreation) {
+  size_t ChunkSize = 256;
+  {
+    NeXusDataset::EventId Cue(RootGroup, NeXusDataset::Mode::Create,
+                               ChunkSize);
+  }
+  ASSERT_TRUE(RootGroup.has_dataset("event_id"));
+  hdf5::node::Dataset TestDataset = RootGroup.get_dataset("event_id");
   auto CreationProperties = TestDataset.creation_list();
   auto ChunkDims = CreationProperties.chunk();
   ASSERT_EQ(ChunkDims.size(), 1u);
   EXPECT_EQ(ChunkDims.at(0), ChunkSize);
-  hdf5::datatype::String StringType(
-      hdf5::datatype::String::fixed(StringLength));
-  StringType.encoding(hdf5::datatype::CharacterEncoding::UTF8);
-  StringType.padding(hdf5::datatype::StringPad::NULLTERM);
-  EXPECT_EQ(StringType, TestDataset.datatype());
+  EXPECT_EQ(hdf5::datatype::create<std::uint32_t>(), TestDataset.datatype());
 }
 
-TEST_F(DatasetCreation, StringDatasetReopen) {
-  std::string DatasetName{"SomeName"};
-  size_t StringLength{24};
-  size_t ChunkSize{511};
-  NeXusDataset::FixedSizeString(RootGroup, DatasetName,
-                                NeXusDataset::Mode::Create, StringLength,
-                                ChunkSize);
-
-  NeXusDataset::FixedSizeString TestDataset(RootGroup, DatasetName,
-                                            NeXusDataset::Mode::Open);
-  EXPECT_EQ(StringLength, TestDataset.getMaxStringSize());
-  EXPECT_EQ(TestDataset.dataspace().size(), 0);
+TEST_F(NeXusDatasetCreation, EventIdReOpen) {
+  {
+    size_t ChunkSize{256};
+    NeXusDataset::EventId Cue(RootGroup, NeXusDataset::Mode::Create,
+                               ChunkSize);
+  }
+  EXPECT_NO_THROW(
+      NeXusDataset::EventId ReOpened(RootGroup, NeXusDataset::Mode::Open));
 }
 
-TEST_F(DatasetCreation, StringDatasetFailReopen) {
-  std::string DatasetName{"SomeName"};
-  EXPECT_THROW(NeXusDataset::FixedSizeString(RootGroup, DatasetName,
-                                             NeXusDataset::Mode::Open),
+TEST_F(NeXusDatasetCreation, EventIdThrowOnExists) {
+  size_t ChunkSize = 256;
+  {
+    NeXusDataset::EventId Cue(RootGroup, NeXusDataset::Mode::Create,
+                               ChunkSize);
+  }
+  EXPECT_THROW(NeXusDataset::EventId Cue(RootGroup, NeXusDataset::Mode::Create,
+                   ChunkSize),
                std::runtime_error);
 }
 
-TEST_F(DatasetCreation, StringDatasetWriteString1) {
-  std::string DatasetName{"SomeName"};
-  size_t StringLength{10};
-  NeXusDataset::FixedSizeString TestDataset(
-      RootGroup, DatasetName, NeXusDataset::Mode::Create, StringLength);
+//--------------------------------------------------
 
-  std::string TestString{"Hello"};
-  TestDataset.appendString(TestString);
-  std::string ReadBackString;
-  TestDataset.read(ReadBackString, TestDataset.datatype(),
-                   hdf5::dataspace::Scalar(),
-                   hdf5::dataspace::Hyperslab{{0}, {1}});
-  std::string CompareString(ReadBackString.data());
-  EXPECT_EQ(TestString, CompareString);
+TEST_F(NeXusDatasetCreation, EventIndexDefaultCreation) {
+  size_t ChunkSize = 256;
+  {
+    NeXusDataset::EventIndex Cue(RootGroup, NeXusDataset::Mode::Create,
+                              ChunkSize);
+  }
+  ASSERT_TRUE(RootGroup.has_dataset("event_index"));
+  hdf5::node::Dataset TestDataset = RootGroup.get_dataset("event_index");
+  auto CreationProperties = TestDataset.creation_list();
+  auto ChunkDims = CreationProperties.chunk();
+  ASSERT_EQ(ChunkDims.size(), 1u);
+  EXPECT_EQ(ChunkDims.at(0), ChunkSize);
+  EXPECT_EQ(hdf5::datatype::create<std::uint32_t>(), TestDataset.datatype());
 }
 
-TEST_F(DatasetCreation, StringDatasetWriteString2) {
-  std::string DatasetName{"SomeName"};
-  size_t StringLength{10};
-  NeXusDataset::FixedSizeString TestDataset(
-      RootGroup, DatasetName, NeXusDataset::Mode::Create, StringLength);
-
-  std::string TestString1{"Hello"};
-  TestDataset.appendString(TestString1);
-  std::string TestString2{"Hi"};
-  TestDataset.appendString(TestString2);
-
-  std::string ReadBackString;
-  TestDataset.read(ReadBackString, TestDataset.datatype(),
-                   hdf5::dataspace::Scalar(),
-                   hdf5::dataspace::Hyperslab{{1}, {1}});
-  std::string CompareString(ReadBackString.data());
-  EXPECT_EQ(TestString2, CompareString);
+TEST_F(NeXusDatasetCreation, EventIndexReOpen) {
+  {
+    size_t ChunkSize{256};
+    NeXusDataset::EventIndex Cue(RootGroup, NeXusDataset::Mode::Create,
+                              ChunkSize);
+  }
+  EXPECT_NO_THROW(
+      NeXusDataset::EventIndex ReOpened(RootGroup, NeXusDataset::Mode::Open));
 }
 
-TEST_F(DatasetCreation, StringDatasetWriteTooLongString) {
-  std::string DatasetName{"SomeName"};
-  size_t StringLength{10};
-  NeXusDataset::FixedSizeString TestDataset(
-      RootGroup, DatasetName, NeXusDataset::Mode::Create, StringLength);
+TEST_F(NeXusDatasetCreation, EventIndexThrowOnExists) {
+  size_t ChunkSize = 256;
+  {
+    NeXusDataset::EventIndex Cue(RootGroup, NeXusDataset::Mode::Create,
+                              ChunkSize);
+  }
+  EXPECT_THROW(NeXusDataset::EventIndex Cue(RootGroup, NeXusDataset::Mode::Create,
+                   ChunkSize),
+               std::runtime_error);
+}
 
-  std::string TestString{"The quick brown fox jumped over the lazy turtle"};
-  TestDataset.appendString(TestString);
-  std::string ReadBackString;
-  TestDataset.read(ReadBackString, TestDataset.datatype(),
-                   hdf5::dataspace::Scalar(),
-                   hdf5::dataspace::Hyperslab{{0}, {1}});
-  std::string CompareString(ReadBackString.data());
-  EXPECT_NE(TestString, CompareString);
-  EXPECT_EQ(std::string(TestString.begin(), TestString.begin() + StringLength),
-            CompareString);
+//--------------------------------------------------
+
+TEST_F(NeXusDatasetCreation, EventTimeOffsetDefaultCreation) {
+  size_t ChunkSize = 256;
+  {
+    NeXusDataset::EventTimeOffset Cue(RootGroup, NeXusDataset::Mode::Create,
+                                 ChunkSize);
+  }
+  ASSERT_TRUE(RootGroup.has_dataset("event_time_offset"));
+  hdf5::node::Dataset TestDataset = RootGroup.get_dataset("event_time_offset");
+  auto CreationProperties = TestDataset.creation_list();
+  auto ChunkDims = CreationProperties.chunk();
+  ASSERT_EQ(ChunkDims.size(), 1u);
+  EXPECT_EQ(ChunkDims.at(0), ChunkSize);
+  EXPECT_EQ(hdf5::datatype::create<std::uint32_t>(), TestDataset.datatype());
+}
+
+TEST_F(NeXusDatasetCreation, EventTimeOffsetReOpen) {
+  {
+    size_t ChunkSize{256};
+    NeXusDataset::EventTimeOffset Cue(RootGroup, NeXusDataset::Mode::Create,
+                                 ChunkSize);
+  }
+  EXPECT_NO_THROW(
+      NeXusDataset::EventTimeOffset ReOpened(RootGroup, NeXusDataset::Mode::Open));
+}
+
+TEST_F(NeXusDatasetCreation, EventTimeOffsetThrowOnExists) {
+  size_t ChunkSize = 256;
+  {
+    NeXusDataset::EventTimeOffset Cue(RootGroup, NeXusDataset::Mode::Create,
+                                 ChunkSize);
+  }
+  EXPECT_THROW(NeXusDataset::EventTimeOffset Cue(RootGroup, NeXusDataset::Mode::Create,
+                   ChunkSize),
+               std::runtime_error);
+}
+
+//--------------------------------------------------
+
+TEST_F(NeXusDatasetCreation, EventTimeZeroDefaultCreation) {
+  size_t ChunkSize = 256;
+  {
+    NeXusDataset::EventTimeZero Cue(RootGroup, NeXusDataset::Mode::Create,
+                                       ChunkSize);
+  }
+  ASSERT_TRUE(RootGroup.has_dataset("event_time_zero"));
+  hdf5::node::Dataset TestDataset = RootGroup.get_dataset("event_time_zero");
+  auto CreationProperties = TestDataset.creation_list();
+  auto ChunkDims = CreationProperties.chunk();
+  ASSERT_EQ(ChunkDims.size(), 1u);
+  EXPECT_EQ(ChunkDims.at(0), ChunkSize);
+  EXPECT_EQ(hdf5::datatype::create<std::uint64_t>(), TestDataset.datatype());
+  bool FoundStartAttr{false};
+  bool FoundUnitAttr{false};
+  for (const auto &Attribute : TestDataset.attributes) {
+    std::string AttributeValue;
+    if (Attribute.name() == "start") {
+      Attribute.read(AttributeValue);
+      if (AttributeValue == "1970-01-01T00:00:00Z") {
+        FoundStartAttr = true;
+      }
+    } else if (Attribute.name() == "units") {
+      Attribute.read(AttributeValue);
+      if (AttributeValue == "ns") {
+        FoundUnitAttr = true;
+      }
+    }
+  }
+  EXPECT_TRUE(FoundStartAttr);
+  EXPECT_TRUE(FoundUnitAttr);
+}
+
+TEST_F(NeXusDatasetCreation, EventTimeZeroReOpen) {
+  {
+    size_t ChunkSize{256};
+    NeXusDataset::EventTimeZero Cue(RootGroup, NeXusDataset::Mode::Create,
+                                       ChunkSize);
+  }
+  EXPECT_NO_THROW(NeXusDataset::EventTimeZero ReOpened(
+                      RootGroup, NeXusDataset::Mode::Open));
+}
+
+TEST_F(NeXusDatasetCreation, EventTimeZeroThrowOnExists) {
+  size_t ChunkSize = 256;
+  {
+    NeXusDataset::EventTimeZero Cue(RootGroup, NeXusDataset::Mode::Create,
+                                       ChunkSize);
+  }
+  EXPECT_THROW(NeXusDataset::EventTimeZero Cue(
+                   RootGroup, NeXusDataset::Mode::Create, ChunkSize),
+               std::runtime_error);
 }
